Apply requested changes round 2
