+++ crates/rust-analyzer/src/reload.rs
@@ -605,7 +605,8 @@ impl GlobalState {
         let mut config_change = ConfigChange::default();
         config_change.change_source_root_parent_map(self.local_roots_parent_map.clone());
         let mut error_sink = ConfigError::default();
-        self.config = Arc::new(self.config.apply_change(config_change, &mut error_sink));
+        let (config, _) = self.config.apply_change(config_change, &mut error_sink);
+        self.config = Arc::new(config);
 
         self.recreate_crate_graph(cause);
